Added (int) to all the sliders so that there won't be any decimals in CharacterCreatePage.xaml

+++ Game/Game/Views/Characters/CharacterCreatePage.xaml.cs
@@ -140,7 +140,7 @@ namespace Game.Views
         /// <param name="e"></param>
         public void Attack_OnSliderValueChanged(object sender, ValueChangedEventArgs e)
         {
-            AttackLabel.Text = String.Format("{0}", e.NewValue);
+            AttackLabel.Text = String.Format("{0}", (int)e.NewValue);
         }
 
         /// <summary>
@@ -150,7 +150,7 @@ namespace Game.Views
         /// <param name="e"></param>
         public void Defense_OnSliderValueChanged(object sender, ValueChangedEventArgs e)
         {
-            DefenseLabel.Text = String.Format("{0}", e.NewValue);
+            DefenseLabel.Text = String.Format("{0}", (int)e.NewValue);
         }
 
         /// <summary>
@@ -160,7 +160,7 @@ namespace Game.Views
         /// <param name="e"></param>
         public void Speed_OnSliderValueChanged(object sender, ValueChangedEventArgs e)
         {
-            SpeedLabel.Text = String.Format("{0}", e.NewValue);
+            SpeedLabel.Text = String.Format("{0}",(int)e.NewValue);
         }
